bump Zephyr to 3.6.0 and zcbor to 0.8.1
Most changes are adjustments to new zcbor API.

Drop NEWLIB_LIBC_FLOAT_PRINTF and PICOLIBC_IO_FLOAT to suppress:

  warning: the choice symbol PICOLIBC_IO_FLOAT (defined at
  lib/libc/picolibc/Kconfig:42) is selected by the following symbols, but
  select/imply has no effect on choice symbols
   - GOLIOTH_FIRMWARE_SDK (defined at
   /modules/lib/golioth-firmware-sdk/port/zephyr/Kconfig:8)

Fix double promotion warnings due to -Wdouble-promotion added to the Zephyr
warning base.

Change zephyr_library_compile_definitions() to
zephyr_compile_definitions(), so that zcbor compatibility macros and
functions are available from application code (like
examples/zephyr/lightdb/get/src/main.c).

Provide zcbor compatibility macros to support both zcbor 0.7.0 (used by
NCS) and zcbor 0.8.1 (used by Zephyr, ESP-IDF and ModusToolbox).

+++ src/fw_block_processor.c
@@ -330,7 +330,7 @@ void fw_block_processor_log_results(const fw_block_processor_ctx_t *ctx)
 {
     GLTH_LOGI(TAG, "Block Latency Stats:");
     GLTH_LOGI(TAG, "   Min: %" PRIu32 " ms", ctx->download.block_stats.block_min_ms);
-    GLTH_LOGI(TAG, "   Ave: %.3f ms", ctx->download.block_stats.block_ema_ms);
+    GLTH_LOGI(TAG, "   Ave: %.3f ms", (double) ctx->download.block_stats.block_ema_ms);
     GLTH_LOGI(TAG, "   Max: %" PRIu32 " ms", ctx->download.block_stats.block_max_ms);
     GLTH_LOGI(TAG, "Total bytes written: %" PRIu32, (uint32_t) ctx->handle_block.bytes_handled);
