+++ src/stream.c
@@ -59,7 +59,7 @@ enum golioth_status golioth_stream_set_float_async(struct golioth_client *client
                                                    void *callback_arg)
 {
     char buf[32] = {};
-    snprintf(buf, sizeof(buf), "%f", value);
+    snprintf(buf, sizeof(buf), "%f", (double) value);
     return golioth_coap_client_set(client,
                                    GOLIOTH_STREAM_PATH_PREFIX,
                                    path,
@@ -173,7 +173,7 @@ enum golioth_status golioth_stream_set_float_sync(struct golioth_client *client,
                                                   int32_t timeout_s)
 {
     char buf[32] = {};
-    snprintf(buf, sizeof(buf), "%f", value);
+    snprintf(buf, sizeof(buf), "%f", (double) value);
     return golioth_coap_client_set(client,
                                    GOLIOTH_STREAM_PATH_PREFIX,
                                    path,
